Break down main into separate files
